【WA】[Kyuride Kagamiz Programming Contest (Easy)]D - 虫歯(Cavity)

+++ k2pc-easy/k2pc001_e4.cpp
@@ -50,7 +50,7 @@ int main() {
 
     set<ll> ope;
 
-    int k, n;
+    ll k, n;
     cin >> k >> n;
 
     map<ll, ll> depth;
